+++ src/main/java/org/folio/dcb/client/feign/CirculationItemClient.java
@@ -7,6 +7,7 @@ import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestParam;
 
@@ -19,4 +20,8 @@ public interface CirculationItemClient {
   CirculationItem retrieveCirculationItemById(@PathVariable("circulationItemId") String circulationItemId);
 
   @GetMapping
-  CirculationItemCollection fetchItemByCqlQuery(@RequestParam("query") String query);}
+  CirculationItemCollection fetchItemByCqlQuery(@RequestParam("query") String query);
+
+  @PutMapping("/{circulationItemId}")
+  CirculationItem updateCirculationItem(@PathVariable("circulationItemId") String circulationItemId, @RequestBody CirculationItem circulationRequest);
+}