+++ src/main/java/org/folio/dcb/service/impl/BorrowingLibraryServiceImpl.java
@@ -33,7 +33,7 @@ public class BorrowingLibraryServiceImpl implements LibraryService {
   private final ServicePointService servicePointService;
   @Override
   public TransactionStatusResponse createCirculation(String dcbTransactionId, DcbTransaction dcbTransaction) {
-    ServicePointRequest pickupServicePoint = servicePointService.createServicePointIfNotExists(dcbTransaction.getPickup());
+    ServicePointRequest pickupServicePoint = servicePointService.createServicePointIfNotExists(dcbTransaction);
     dcbTransaction.getPickup().setServicePointId(pickupServicePoint.getId());
     return libraryService.createBorrowingLibraryTransaction(dcbTransactionId, dcbTransaction, pickupServicePoint.getId());
   }