+++ src/main/java/org/folio/dcb/utils/JsonUtils.java
@@ -1,6 +1,7 @@
 package org.folio.dcb.utils;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import lombok.experimental.UtilityClass;
 import lombok.extern.log4j.Log4j2;
@@ -20,4 +21,19 @@ public class JsonUtils {
     }
   }
 
+  public static JsonNode readTree(String value) {
+    try {
+      return OBJECT_MAPPER.readTree(value);
+    } catch (JsonProcessingException e) {
+      throw new RuntimeException("Failed to read string value", e);
+    }
+  }
+
+  public static String writeValueAsString(Object value) {
+    try {
+      return OBJECT_MAPPER.writeValueAsString(value);
+    } catch (JsonProcessingException e) {
+      throw new IllegalArgumentException("Failed to convert value to JSON string", e);
+    }
+  }
 }