+++ src/test/java/org/folio/dcb/it/BorrowingPickupTransactionIT.java
@@ -9,6 +9,7 @@ import static com.github.tomakehurst.wiremock.client.WireMock.urlPathEqualTo;
 import static com.github.tomakehurst.wiremock.client.WireMock.urlPathMatching;
 import static java.lang.String.format;
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.folio.dcb.domain.dto.TransactionStatus.StatusEnum.AWAITING_PICKUP;
 import static org.folio.dcb.domain.dto.TransactionStatus.StatusEnum.CANCELLED;
 import static org.folio.dcb.domain.dto.TransactionStatus.StatusEnum.CLOSED;
 import static org.folio.dcb.domain.dto.TransactionStatus.StatusEnum.CREATED;
@@ -16,7 +17,7 @@ import static org.folio.dcb.domain.dto.TransactionStatus.StatusEnum.EXPIRED;
 import static org.folio.dcb.domain.dto.TransactionStatus.StatusEnum.ITEM_CHECKED_IN;
 import static org.folio.dcb.domain.dto.TransactionStatus.StatusEnum.ITEM_CHECKED_OUT;
 import static org.folio.dcb.domain.dto.TransactionStatus.StatusEnum.OPEN;
-import static org.folio.dcb.utils.EntityUtils.BORROWER_SERVICE_POINT_ID;
+import static org.folio.dcb.utils.EntityUtils.VIRTUAL_SERVICE_POINT_ID;
 import static org.folio.dcb.utils.EntityUtils.DCB_TRANSACTION_ID;
 import static org.folio.dcb.utils.EntityUtils.EXISTED_PATRON_ID;
 import static org.folio.dcb.utils.EntityUtils.ITEM_ID;
@@ -24,12 +25,14 @@ import static org.folio.dcb.utils.EntityUtils.LOAN_ID;
 import static org.folio.dcb.utils.EntityUtils.NOT_EXISTED_PATRON_ID;
 import static org.folio.dcb.utils.EntityUtils.PATRON_TYPE_USER_ID;
 import static org.folio.dcb.utils.EntityUtils.PICKUP_SERVICE_POINT_ID;
-import static org.folio.dcb.utils.EntityUtils.borrowerDcbTransaction;
+import static org.folio.dcb.utils.EntityUtils.TEST_TENANT;
 import static org.folio.dcb.utils.EntityUtils.borrowingPickupDcbTransaction;
 import static org.folio.dcb.utils.EntityUtils.dcbItem;
 import static org.folio.dcb.utils.EntityUtils.dcbPatron;
 import static org.folio.dcb.utils.EntityUtils.dcbTransactionUpdate;
 import static org.folio.dcb.utils.EntityUtils.transactionStatus;
+import static org.folio.dcb.utils.EventDataProvider.expiredRequestMessage;
+import static org.folio.dcb.utils.EventDataProvider.itemCheckInMessage;
 import static org.folio.dcb.utils.JsonTestUtils.asJsonString;
 import static org.hamcrest.Matchers.containsString;
 import static org.hamcrest.Matchers.is;
@@ -284,7 +287,7 @@ class BorrowingPickupTransactionIT extends BaseTenantIntegrationTest {
       .withRequestBody(matchingJsonPath("$.status", equalTo("Closed - Cancelled")))
       .withRequestBody(matchingJsonPath("$.instanceId", equalTo(DCBConstants.INSTANCE_ID)))
       .withRequestBody(matchingJsonPath("$.requesterId", equalTo(PATRON_TYPE_USER_ID)))
-      .withRequestBody(matchingJsonPath("$.pickupServicePointId", equalTo(BORROWER_SERVICE_POINT_ID)))
+      .withRequestBody(matchingJsonPath("$.pickupServicePointId", equalTo(VIRTUAL_SERVICE_POINT_ID)))
       .withRequestBody(matchingJsonPath("$.holdingsRecordId", equalTo(DCBConstants.HOLDING_ID))));
   }
 
@@ -302,13 +305,36 @@ class BorrowingPickupTransactionIT extends BaseTenantIntegrationTest {
   @EnumSource(value = TransactionStatus.StatusEnum.class, names = "EXPIRED", mode = EXCLUDE)
   void updateTransactionStatus_parameterized_invalidTransitionToExpiredStatus(
     TransactionStatus.StatusEnum sourceStatus) throws Exception {
-    testJdbcHelper.saveDcbTransaction(DCB_TRANSACTION_ID, sourceStatus, borrowerDcbTransaction());
+    testJdbcHelper.saveDcbTransaction(DCB_TRANSACTION_ID, sourceStatus, borrowingPickupDcbTransaction());
 
     putDcbTransactionStatusAttempt(DCB_TRANSACTION_ID, transactionStatus(EXPIRED))
       .andExpect(status().isBadRequest())
       .andExpect(jsonPath("$.errors[0].code").value("VALIDATION_ERROR"))
       .andExpect(jsonPath("$.errors[0].message").value(containsString(String.format(
         "Status transition will not be possible from %s to EXPIRED", sourceStatus))));
+  }
+
+  @Test
+  void updateStatus_positive_awaitingPickupTransactionExpiration() throws Exception {
+    testJdbcHelper.saveDcbTransaction(DCB_TRANSACTION_ID, AWAITING_PICKUP, borrowingPickupDcbTransaction());
+    getDcbTransactionStatus(DCB_TRANSACTION_ID)
+      .andExpect(jsonPath("$.status").value(AWAITING_PICKUP.getValue()));
+
+    testEventHelper.sendMessage(expiredRequestMessage(TEST_TENANT));
+    awaitUntilAsserted(() -> getDcbTransactionStatus(DCB_TRANSACTION_ID)
+      .andExpect(jsonPath("$.status").value(EXPIRED.getValue())));
+  }
+
+  @Test
+  void updateStatus_positive_expiredToClosedTransitionAfterCheckInMessage() throws Exception {
+    testJdbcHelper.saveDcbTransaction(DCB_TRANSACTION_ID, EXPIRED, borrowingPickupDcbTransaction());
+    getDcbTransactionStatus(DCB_TRANSACTION_ID)
+      .andExpect(jsonPath("$.status").value(EXPIRED.getValue()));
+
+    testEventHelper.sendMessage(itemCheckInMessage(TEST_TENANT));
+
+    awaitUntilAsserted(() -> getDcbTransactionStatus(DCB_TRANSACTION_ID)
+      .andExpect(jsonPath("$.status").value(CLOSED.getValue())));
   }
 
   private static void verifyPostCirculationRequestCalledOnce(String requesterId) {